factored out <ModalAddOrChangeYourClub> and <ModalRemoveYourClub>; created a /clubs/list page
