+++ less/components/hero-unit.less
@@ -45,7 +45,7 @@
 
   h2 {
     margin: 0;
-    font-size: 3rem;
+    font-size: 2.6rem;
     color: @white;
   }
