+++ less/components/sidebar.less
@@ -24,6 +24,7 @@ html.no-js .sidebar {
     .sidebarColorizer(@greenSidebarColor);
   }
   &.clubs,
+  &.clubs-list,
   &.clubs-toolkit {
     .sidebarColorizer(@orangeSidebarColor);
   }
