+++ less/pages/clubs-list.less
@@ -1,0 +1,11 @@
+.page.clubs-list .hero-unit {
+  background-image: url(/img/pages/clubs/hero-clubs.jpg);
+
+  .retina2x({
+    background-image: url(/img/pages/clubs/hero-clubs@2x.jpg);
+  });
+
+  .btn {
+    margin-top: 7rem;
+  }
+}
