+++ less/pages/index.less
@@ -8,6 +8,7 @@
 @import "events";
 @import "event-resources";
 @import "web-literacy";
+@import "clubs-list";
 @import "clubs-toolkit";
 @import "tools";
 @import "madewithcode";
