+++ lib/routes.jsx
@@ -35,6 +35,8 @@ var routes = (
     <Redirect from="/clubs/curriculum/" to="/activities/web-lit-basics/" />
     <Route name="mozilla-clubs" path="/clubs/"
      handler={require('../pages/clubs.jsx')}/>
+    <Route name="clubs-list" path="/clubs/list/"
+     handler={require('../pages/clubs-list.jsx')}/>
     <Route name="clubs-toolkit" path="/clubs/toolkit/"
      handler={require('../pages/clubs-toolkit.jsx')}/>
     <Route name="events" path="/events/"
